Create an update option to install faster
The option is not yet propagated, but it is there. Installs are
now checking if the dependencies are there, and only install if
necessary, unless the update option is supplied.

Also a lot of small improvements added, as well as more logs.

Also added test-run-docker.sh that runs a docker container
while mounting the scripts, making it easier to iterate when
testing in a container.