Fix starship not showing exit code
This was coming from a conflict between kitty and history -a.
When a process was pending in a tab, if you created a new tab
this new tab would no longer show the exit status code.
By moving it to the end of PROMPT_COMMAND we fix this.
Did this by disabling kitty auto shell integration, which is
also added twice - another conflict with history -a - and
by moving history -a to the end.